Corrected File Production in V1DD Wrangling Script
I applied the discussed edits to reorder the assemblies by their size, as we do in the figures. I also changed the functional indexes from being 1-indexed (which is from the MATLAB original application) to a 0-index. I also changed it so that the assemblies are now 1-indexed for ease-of-plotting purposes. All changes are commented with the flag "### JULIAN EDIT".